Move settings into last nav bar button. Move favorites to third button of nav. #24
