+++ src/application/controller/Settings/settings.css
@@ -9,7 +9,7 @@
     -fx-text-fill: #F2F2F2;
 }
 
-#playlistsLabel {
+#settingsLabel {
     -fx-font-family: "Verdana";
     -fx-font-weight: bold;
     -fx-font-size: 22;
